+++ node/src/GlideClusterClient.ts
@@ -34,6 +34,7 @@ import {
     createFunctionList,
     createFunctionLoad,
     createInfo,
+    createLastSave,
     createLolwut,
     createPing,
     createPublish,
@@ -934,7 +935,7 @@ export class GlideClusterClient extends BaseClient {
      *
      * See https://valkey.io/commands/dbsize/ for more details.
 
-     * @param route - The command will be routed to all primaries, unless `route` is provided, in which
+     * @param route - The command will be routed to all primary nodes, unless `route` is provided, in which
      *     case the client will route the command to the nodes defined by `route`.
      * @returns The number of keys in the database.
      *     In the case of routing the query to multiple nodes, returns the aggregated number of keys across the different nodes.
@@ -945,7 +946,7 @@ export class GlideClusterClient extends BaseClient {
      * console.log("Number of keys across all primary nodes: ", numKeys);
      * ```
      */
-    public dbsize(route?: Routes): Promise<ClusterResponse<number>> {
+    public dbsize(route?: Routes): Promise<number> {
         return this.createWritePromise(createDBSize(), toProtobufRoute(route));
     }
 
@@ -981,6 +982,28 @@ export class GlideClusterClient extends BaseClient {
     ): Promise<number> {
         return this.createWritePromise(
             createPublish(message, channel, sharded),
+        );
+    }
+
+    /**
+     * Returns `UNIX TIME` of the last DB save timestamp or startup timestamp if no save
+     * was made since then.
+     *
+     * See https://redis.io/commands/lastsave/ for more details.
+     *
+     * @param route - The command will be routed to a random node, unless `route` is provided, in which
+     *     case the client will route the command to the nodes defined by `route`.
+     * @returns `UNIX TIME` of the last DB save executed with success.
+     * @example
+     * ```typescript
+     * const timestamp = await client.lastsave();
+     * console.log("Last DB save was done at " + timestamp);
+     * ```
+     */
+    public async lastsave(route?: Routes): Promise<ClusterResponse<number>> {
+        return this.createWritePromise(
+            createLastSave(),
+            toProtobufRoute(route),
         );
     }
 }