+++ node/tests/RedisClient.test.ts
@@ -12,7 +12,7 @@ import {
 } from "@jest/globals";
 import { BufferReader, BufferWriter } from "protobufjs";
 import { v4 as uuidv4 } from "uuid";
-import { GlideClient, ProtocolVersion, Transaction } from "..";
+import { GlideClient, ListDirection, ProtocolVersion, Transaction } from "..";
 import { RedisCluster } from "../../utils/TestUtils.js";
 import { FlushMode } from "../build-ts/src/Commands";
 import { command_request } from "../src/ProtobufMessage";
@@ -29,7 +29,6 @@ import {
     transactionTest,
     validateTransactionResponse,
 } from "./TestUtilities";
-import { ListDirection } from "..";
 
 /* eslint-disable @typescript-eslint/no-var-requires */
 
@@ -666,6 +665,26 @@ describe("GlideClient", () => {
                 expect(await client.functionFlush()).toEqual("OK");
                 client.close();
             }
+        },
+    );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        "lastsave %p",
+        async (protocol) => {
+            const client = await GlideClient.createClient(
+                getClientConfigurationOption(cluster.getAddresses(), protocol),
+            );
+
+            const today = new Date();
+            today.setDate(today.getDate() - 1);
+            const yesterday = today.getTime() / 1000; // as epoch time
+
+            expect(await client.lastsave()).toBeGreaterThan(yesterday);
+
+            const response = await client.exec(new Transaction().lastsave());
+            expect(response?.[0]).toBeGreaterThan(yesterday);
+
+            client.close();
         },
     );