+++ node/tests/RedisClusterClient.test.ts
@@ -984,4 +984,29 @@ describe("GlideClusterClient", () => {
             );
         },
     );
+
+    it.each([ProtocolVersion.RESP2, ProtocolVersion.RESP3])(
+        "lastsave %p",
+        async (protocol) => {
+            const client = await GlideClusterClient.createClient(
+                getClientConfigurationOption(cluster.getAddresses(), protocol),
+            );
+
+            const today = new Date();
+            today.setDate(today.getDate() - 1);
+            const yesterday = today.getTime() / 1000; // as epoch time
+
+            expect(await client.lastsave()).toBeGreaterThan(yesterday);
+            Object.values(await client.lastsave("allNodes")).forEach((v) =>
+                expect(v).toBeGreaterThan(yesterday),
+            );
+
+            const response = await client.exec(
+                new ClusterTransaction().lastsave(),
+            );
+            expect(response?.[0]).toBeGreaterThan(yesterday);
+
+            client.close();
+        },
+    );
 });